Cover more cases with AutoFixAttributes

+++ tools/FreecraftCore.Tools.AutoFixAttributes/Program.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using FreecraftCore.Serializer;
 using Microsoft.CodeAnalysis;
 
 namespace FreecraftCore
@@ -12,9 +12,12 @@ namespace FreecraftCore
 	{
 		static async Task Main(string[] args)
 		{
-			foreach (string filePath in Directory.GetFiles(Directory.GetCurrentDirectory(), "*.cs"))
+			foreach (string filePath in Directory.GetFiles(Directory.GetCurrentDirectory(), "*.cs", SearchOption.AllDirectories))
 			{
 				string[] file = await File.ReadAllLinesAsync(filePath);
+
+				bool isFileModified = false;
+				int errorCount = 0;
 
 				//Find the a WireMember
 				for (int i = 0; i < file.Length; i++)
@@ -26,19 +29,38 @@ namespace FreecraftCore
 						{
 							//TODO: Support same line.
 							//Attributes are done.
-							if (file[j][0] != '[')
+							if (file[j].Contains("{ get"))
 							{
+								string original = file[j];
+
 								file[j] = file[j].Replace("{ get; }", "{ get; internal set; }");
-								file[j] = file[j].Replace("private set", "internal set");
+
+								//All privates should be internal
+								file[j] = file[j].Replace("private ", "internal ");
 								file[j] = file[j].Replace("protected set", "internal set");
 								file[j] = file[j].Replace("{ get; set; }", "{ get; internal set; }");
+
+								//Cover case where it's internal i { get; internal set; }
+								if (file[j].Contains("internal set") && file[j].Contains("\tinternal "))
+									file[j] = file[j].Replace("internal set", "set");
+
+								if (original != file[j])
+								{
+									isFileModified = true;
+									errorCount++;
+								}
+
 								break;
 							}
 						}
 					}
 				}
 
-				await File.WriteAllLinesAsync(filePath, file);
+				if (isFileModified)
+				{
+					Console.WriteLine($"Fixed Errors in {Path.GetFileNameWithoutExtension(filePath)}: {errorCount}");
+					await File.WriteAllLinesAsync(filePath, file);
+				}
 			}
 		}
 	}
